# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Djobea Analytics API",
    "description": "API complète pour la plateforme de gestion de services à domicile Djobea",
    "version": "1.0.0",
    "contact": {
      "name": "Équipe Djobea",
      "email": "support@djobea.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://djobea-analytics.vercel.app/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/providers": {
      "get": {
        "tags": ["Providers"],
        "summary": "Lister les prestataires",
        "description": "Récupère la liste paginée des prestataires avec filtres",
        "responses": {
          "200": {
            "description": "Liste des prestataires récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "providers": { "type": "array" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Providers"],
        "summary": "Créer un nouveau prestataire",
        "description": "Ajoute un nouveau prestataire au système",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prestataire créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
    "/providers/{id}": {
      "get": {
        "tags": ["Providers"],
        "summary": "Récupérer un prestataire spécifique",
        "description": "Retourne les détails d'un prestataire par son ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID du prestataire",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prestataire récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Prestataire non trouvé"
          }
        }
      },
      "put": {
        "tags": ["Providers"],
        "summary": "Mettre à jour un prestataire",
        "description": "Met à jour les informations d'un prestataire",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID du prestataire",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prestataire mis à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Prestataire non trouvé"
          },
          "400": {
            "description": "Données invalides"
          }
        }
      },
      "delete": {
        "tags": ["Providers"],
        "summary": "Supprimer un prestataire",
        "description": "Supprime un prestataire du système",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID du prestataire",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prestataire supprimé avec succès"
          },
          "404": {
            "description": "Prestataire non trouvé"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Provider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          }
        }
      },
...
      "HourlyRate": {
        "type": "object",
        "properties": {
          "min": {
            "type": "number"
          },
          "max": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "Availability": {
        "type": "object",
        "properties": {
          "monday": {
            "$ref": "#/components/schemas/TimeSlot"
          },
          "tuesday": {
            "$ref": "#/components/schemas/TimeSlot"
          },
          "wednesday": {
            "$ref": "#/components/schemas/TimeSlot"
          },
          "thursday": {
            "$ref": "#/components/schemas/TimeSlot"
          },
          "friday": {
            "$ref": "#/components/schemas/TimeSlot"
          },
          "saturday": {
            "$ref": "#/components/schemas/TimeSlot"
          },
          "sunday": {
            "$ref": "#/components/schemas/TimeSlot"
          }
        }
      },
      "TimeSlot": {
        "type": "object",
        "nullable": true,
        "properties": {
          "start": {
            "type": "string",
            "format": "time"
          },
          "end": {
            "type": "string",
            "format": "time"
          }
        }
      },
...
      "CreateProviderInput": {
        "type": "object",
        "required": ["name", "email", "phone", "specialty", "zone"],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "certifications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hourlyRate": {
            "$ref": "#/components/schemas/HourlyRate"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          }
        }
      },
      "UpdateProviderInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "pending"]
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "certifications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hourlyRate": {
            "$ref": "#/components/schemas/HourlyRate"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          }
        }
      },
...
      "ProviderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Provider"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ]
}